requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-binance==1.0.19
supabase==2.3.0
structlog==24.1.0
httpx==0.26.0
python-multipart==0.0.6
